Refactored aurite studio codebase: Implement sidebar navigation and main pages for agent and workflow management as routes & modular:
- Added Sidebar component for navigation with dynamic item highlighting and tooltips.
- Created AgentsPage for managing agents with options to create, edit, and run agents.
- Developed HomePage for creating new agents with a guided interface and quick action suggestions.
- Implemented LLMConfigsPage for managing language model configurations.
- Added MCPClientsPage for managing MCP client connections.
- Created WorkflowsPage for designing and managing agent workflows with execution capabilities.
- Integrated loading states and empty states for better user experience across pages.
